+++ CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Logo
 - `Response` class `top_k` property
+- `Response` class `top_p` property
 ### Changed
 - `AI_STUDIO` render format modified
 - Test system modified
+- `README.md` updated
 ## [0.7] - 2025-06-25
 ### Added
 - `Message` abstract class